# Span Symmetric ICryptoTransform