--- a/src/libraries/Common/src/Internal/Cryptography/BasicSymmetricCipher.cs
+++ b/src/libraries/Common/src/Internal/Cryptography/BasicSymmetricCipher.cs
@@ -31,7 +31,7 @@ namespace Internal.Cryptography
 
         public abstract int Transform(byte[] input, int inputOffset, int count, byte[] output, int outputOffset);
 
-        public abstract byte[] TransformFinal(byte[] input, int inputOffset, int count);
+        public abstract int TransformFinal(ReadOnlySpan<byte> input, Span<byte> output);
 
         public int BlockSizeInBytes { get; private set; }