--- a/src/libraries/System.Security.Cryptography.Algorithms/src/Internal/Cryptography/AppleCCCryptor.cs
+++ b/src/libraries/System.Security.Cryptography.Algorithms/src/Internal/Cryptography/AppleCCCryptor.cs
@@ -68,6 +68,8 @@ namespace Internal.Cryptography
             int ret;
             int errorCode;
 
+            Debug.Assert(output.Length > 0);
+
             fixed (byte* outputStart = output)
             {
                 byte* outputCurrent = outputStart + outputBytes;