… src/libraries/Common/src/Internal/Cryptography/UniversalCryptoTransform.cs
… src/libraries/Common/src/Interop/Windows/BCrypt/Interop.BCryptEncryptDecrypt.cs
… src/libraries/System.Security.Cryptography.Algorithms/src/Internal/Cryptography/OpenSslCipher.cs
… src/libraries/System.Security.Cryptography.Cng/src/Internal/Cryptography/BasicSymmetricCipherNCrypt.cs
… src/libraries/System.Security.Cryptography.Cng/src/Resources/Strings.resx
… src/libraries/System.Security.Cryptography.Csp/src/Internal/Cryptography/BasicSymmetricCipherCsp.cs
… src/libraries/System.Security.Cryptography.Csp/src/Resources/Strings.resx
… src/libraries/System.Security.Cryptography.Csp/src/System.Security.Cryptography.Csp.csproj
… src/libraries/System.Security.Cryptography.Csp/src/System/Security/Cryptography/CapiHelper.Windows.cs