This changes the underlying `UniversalCryptoTransform`, derived types, and symmetric cryptographic primitives to operate on span.

This gains some performance benefits already, either by avoiding temporary arrays (removing and adding padding) and using `CryptoPool` where possible.

~This also would pave the way for `{ReadOnly}Span<byte>` to `ICryptoTransform` and the built-in primitives, provided that a sensible default behavior is possible with DIM.~

These changes are also motivated by #2406, which will operate on span.

<details>
<summary>Performance Numbers</summary>
<pre>
|                      Method |        Job |   Toolchain | DataSize |  Mode |            Mean |         Error |        StdDev |          Median | Ratio | RatioSD |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|---------------------------- |----------- |------------ |--------- |------ |----------------:|--------------:|--------------:|----------------:|------:|--------:|---------:|---------:|---------:|----------:|
|      Encrypt_TransformBlock | Job-QQUEGF | crypto-span |  1048576 | PKCS7 | 1,026,981.62 ns |  9,912.516 ns |  8,787.188 ns | 1,029,210.30 ns |  1.00 |    0.01 |        - |        - |        - |       1 B |
|      Encrypt_TransformBlock | Job-PLPBZT |      master |  1048576 | PKCS7 | 1,024,755.20 ns |  3,505.305 ns |  3,278.864 ns | 1,024,598.03 ns |  1.00 |    0.00 |        - |        - |        - |       1 B |
|                             |            |             |          |       |                 |               |               |                 |       |         |          |          |          |           |
|      Decrypt_TransformBlock | Job-QQUEGF | crypto-span |  1048576 | PKCS7 |   135,297.81 ns |    453.515 ns |    424.218 ns |   135,393.02 ns |  0.98 |    0.01 |        - |        - |        - |         - |
|      Decrypt_TransformBlock | Job-PLPBZT |      master |  1048576 | PKCS7 |   138,163.82 ns |    687.527 ns |    609.475 ns |   138,280.70 ns |  1.00 |    0.00 |        - |        - |        - |         - |
|                             |            |             |          |       |                 |               |               |                 |       |         |          |          |          |           |
| Encrypt_TransformFinalBlock | Job-QQUEGF | crypto-span |  1048576 | PKCS7 | 1,225,488.74 ns | 11,252.529 ns |  9,975.074 ns | 1,226,682.06 ns |  0.80 |    0.01 | 195.3125 | 195.3125 | 195.3125 | 1048668 B |
| Encrypt_TransformFinalBlock | Job-PLPBZT |      master |  1048576 | PKCS7 | 1,531,450.18 ns | 12,607.728 ns | 10,528.024 ns | 1,534,052.30 ns |  1.00 |    0.00 | 250.0000 | 250.0000 | 250.0000 | 2097299 B |
|                             |            |             |          |       |                 |               |               |                 |       |         |          |          |          |           |
| Decrypt_TransformFinalBlock | Job-QQUEGF | crypto-span |  1048576 | PKCS7 |   806,700.19 ns | 11,034.824 ns | 10,321.981 ns |   808,663.70 ns |  0.73 |    0.03 | 397.4609 | 397.4609 | 397.4609 | 3145944 B |
| Decrypt_TransformFinalBlock | Job-PLPBZT |      master |  1048576 | PKCS7 | 1,111,906.51 ns | 21,502.634 ns | 27,193.948 ns | 1,121,153.26 ns |  1.00 |    0.00 | 269.5313 | 269.5313 | 269.5313 | 3145905 B |

</pre>
</details>