@vcsjones 2020-07-03
Presumably the AllocateUnitializedArray calls need a `#if NETCOREAPP5_0` guard and just new-up a regular array otherwise. Will address.