@vcsjones 2020-07-10
> added a few nitpicky things that I was otherwise going to just be quiet about.

I'm always happy to fix things!